+++ bachelor/src/main/java/com/erim/bachelor/controller/AuthenticationController.java
@@ -1,11 +1,10 @@
 package com.erim.bachelor.controller;
 
-import com.erim.bachelor.data.AuthenticationRequest;
-import com.erim.bachelor.data.AuthenticationResponse;
-import com.erim.bachelor.data.PasswordChangeDTO;
+import com.erim.bachelor.dto.AuthenticationRequest;
+import com.erim.bachelor.dto.AuthenticationResponse;
+import com.erim.bachelor.dto.PasswordChangeDTO;
 import com.erim.bachelor.security.AuthenticationService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
