+++ bachelor/src/main/java/com/erim/bachelor/dto/PasswordChangeDTO.java
@@ -1,4 +1,4 @@
-package com.erim.bachelor.data;
+package com.erim.bachelor.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
