+++ bachelor/src/main/java/com/erim/bachelor/security/AuthenticationService.java
@@ -1,6 +1,6 @@
 package com.erim.bachelor.security;
 
-import com.erim.bachelor.data.*;
+import com.erim.bachelor.dto.*;
 import com.erim.bachelor.entities.Borrower;
 import com.erim.bachelor.enums.BorrowerState;
 import com.erim.bachelor.enums.Role;
